+++ agent/packaging/deb/mkdeb.sh
@@ -66,15 +66,6 @@ rm -vf $(find "$DESTDIR/usr/lib" -name "*cpython-*.pyc" | grep -v "cpython-${pyt
 
 #       B U I L D
 
-fpm_args=()
-case "$codename" in
-	stretch)
-	;;
-	*)
-		fpm_args+=(--depends python3-distutils)
-	;;
-esac
-
 fpm --verbose \
     --force \
     --debug-workspace \
@@ -96,6 +87,7 @@ fpm --verbose \
     --depends python3-pkg-resources \
     --depends 'python3-psycopg2 >= 2.7' \
     --depends python3-setuptools \
+    --depends python3-distutils \
     --depends ssl-cert \
     --after-install share/restart-all.sh \
     "$@" \
@@ -108,11 +100,6 @@ mv "$deb" dist/
 dpkg-deb --info "dist/$deb"
 dpkg-deb --show --showformat '$''{Depends}\n' "dist/$deb"
 dpkg-deb --contents "dist/$deb"
-if grep -q stretch /etc/os-release ; then
-	# Debian has only python3-psycopg2 2.6. Use python3-psycopg2 >2.7 from PGDG.
-	curl https://www.postgresql.org/media/keys/ACCC4CF8.asc | gpg --dearmor > /etc/apt/trusted.gpg.d/apt.postgresql.org.gpg
-	echo "deb http://apt-archive.postgresql.org/pub/repos/apt $(lsb_release -cs)-pgdg main" > /etc/apt/sources.list.d/pgdg.list
-fi
 apt-get update --quiet
 apt-get install --yes "./dist/$deb"
 (